For #6261 - Move browser.menu2 to concept-menu2
Move the current dumb data classes into a concept component
