+++ .buildconfig.yml
@@ -20,6 +20,10 @@ projects:
     path: components/concept/engine
     description: 'An abstract layer hiding the actual browser engine implementation.'
     publish: true
+  concept-menu:
+    path: components/concept/menu
+    description: 'An abstract definition of a browser menu component.'
+    publish: true
   concept-push:
     path: components/concept/push
     description: 'An abstract definition of a push service component.'
