+++ components/concept/menu/build.gradle
@@ -22,27 +22,16 @@ android {
 }
 
 dependencies {
-    implementation project(':concept-menu')
-    implementation project(':support-base')
-    implementation project(':support-ktx')
-    implementation project(':ui-icons')
-
-    implementation Dependencies.androidx_appcompat
-    implementation Dependencies.androidx_core_ktx
-    implementation Dependencies.androidx_recyclerview
-    implementation Dependencies.androidx_cardview
-    implementation Dependencies.androidx_constraintlayout
-
     implementation Dependencies.kotlin_stdlib
 
-    testImplementation project(':support-test')
+    implementation Dependencies.androidx_annotation
+    implementation Dependencies.androidx_appcompat
+    implementation project(':support-ktx')
 
-    testImplementation Dependencies.androidx_test_core
     testImplementation Dependencies.androidx_test_junit
     testImplementation Dependencies.testing_robolectric
     testImplementation Dependencies.testing_mockito
-    testImplementation Dependencies.testing_coroutines
-
+    testImplementation project(':support-test')
 }
 
 apply from: '../../../publish.gradle'
