+++ components/concept/menu/src/main/java/mozilla/components/concept/menu/Side.kt
@@ -1,0 +1,19 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package mozilla.components.concept.menu
+
+/**
+ * Indicates the starting or ending side of the menu or an option.
+ */
+enum class Side {
+    /**
+     * Starting side (top or left).
+     */
+    START,
+    /**
+     * Ending side (bottom or right).
+     */
+    END
+}
