+++ components/concept/menu/src/main/java/mozilla/components/concept/menu/candidate/TextStyle.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.browser.menu2.candidate
+package mozilla.components.concept.menu.candidate
 
 import android.graphics.Typeface
 import android.view.View
